Added an :except option to default_actions helper method on IndexAsTable. Excepts any of the current three actions it builds links for: :view, :edit, and :delete

+++ lib/active_admin/views/index_as_table.rb
@@ -33,11 +33,13 @@ module ActiveAdmin
         def default_actions(options = {})
           options = {
             :name => ""
+            :except => []
           }.merge(options)
           column options[:name] do |resource|
-            links = link_to "View", resource_path(resource), :class => "view_link"
-            links += link_to "Edit", edit_resource_path(resource), :class => "edit_link"
-            links += link_to "Delete", resource_path(resource), :method => :delete, :confirm => "Are you sure you want to delete this?", :class => "delete_link"
+            links = ''
+            links += link_to "View", resource_path(resource), :class => "view_link" unless options[:except].include?(:view)
+            links += link_to "Edit", edit_resource_path(resource), :class => "edit_link" unless options[:except].include?(:edit)
+            links += link_to "Delete", resource_path(resource), :method => :delete, :confirm => "Are you sure you want to delete this?", :class => "delete_link" unless options[:except].include?(:delete)
             links
           end
         end
